API-19: make sure that only the owner can delete

+++ app/Http/Controllers/Question/DeleteController.php
@@ -9,6 +9,8 @@ class DeleteController extends Controller
 {
     public function __invoke(Question $question)
     {
+        $this->authorize('forceDelete', $question);
+
         $question->forceDelete();
 
         return response()->noContent();
